+++ tests/Sabre/CalDAV/Schedule/IMipPluginTest.php
@@ -11,9 +11,9 @@ class IMipPluginTest extends \PHPUnit_Framework_TestCase {
     function testDeliverReply() {
 
         $message = new Message();
-        $message->sender = 'sender@example.org';
+        $message->sender = 'mailto:sender@example.org';
         $message->senderName = 'Sender';
-        $message->recipient = 'recipient@example.org';
+        $message->recipient = 'mailto:recipient@example.org';
         $message->recipientName = 'Recipient';
         $message->method = 'REPLY';
 
@@ -50,12 +50,42 @@ ICS;
 
     }
 
+    function testDeliverReplyNoMailto() {
+
+        $message = new Message();
+        $message->sender = 'mailto:sender@example.org';
+        $message->senderName = 'Sender';
+        $message->recipient = 'http://example.org/recipient';
+        $message->recipientName = 'Recipient';
+        $message->method = 'REPLY';
+
+        $ics = <<<ICS
+BEGIN:VCALENDAR\r
+METHOD:REPLY\r
+BEGIN:VEVENT\r
+SUMMARY:Birthday party\r
+END:VEVENT\r
+END:VCALENDAR\r
+
+ICS;
+
+
+        $message->message = Reader::read($ics);
+
+        $result = $this->schedule($message);
+
+        $expected = [];
+
+        $this->assertEquals($expected, $result);
+
+    }
+
     function testDeliverRequest() {
 
         $message = new Message();
-        $message->sender = 'sender@example.org';
+        $message->sender = 'mailto:sender@example.org';
         $message->senderName = 'Sender';
-        $message->recipient = 'recipient@example.org';
+        $message->recipient = 'mailto:recipient@example.org';
         $message->recipientName = 'Recipient';
         $message->method = 'REQUEST';
 
@@ -95,9 +125,9 @@ ICS;
     function testDeliverCancel() {
 
         $message = new Message();
-        $message->sender = 'sender@example.org';
+        $message->sender = 'mailto:sender@example.org';
         $message->senderName = 'Sender';
-        $message->recipient = 'recipient@example.org';
+        $message->recipient = 'mailto:recipient@example.org';
         $message->recipientName = 'Recipient';
         $message->method = 'CANCEL';
 
@@ -150,9 +180,9 @@ ICS;
     function testDeliverInsignificantRequest() {
 
         $message = new Message();
-        $message->sender = 'sender@example.org';
+        $message->sender = 'mailto:sender@example.org';
         $message->senderName = 'Sender';
-        $message->recipient = 'recipient@example.org';
+        $message->recipient = 'mailto:recipient@example.org';
         $message->recipientName = 'Recipient';
         $message->method = 'REQUEST';
         $message->significantChange = false;